Main-/createDoc(()/ createDoc ()- (voorraadbeheer)-04/04/2024: This function take a name of an item, a number of item in total and an array to construct a table. Searches a file in the drive, check to see if the given foldername exist in the drive, if not the script creates it, then moves the file in the folder.
Main-/createDoc(()/ createDoc ()- (voorraadbeheer)-04/04/2024: This function take a name of an item, a number of item in total and an array to construct a table. Searches a file in the drive, check to see if the given foldername exist in the drive, if not the script creates it, then moves the file in the folder.