Main- itemPresentInList()/ minimumSupply()- (voorraadbeheer)-09/04/2024: MinimumSupply function checks all the items in the supply and calculates the total supply and compares these with the minimum supply necessary. If below a treshold adds them to the list if not already present.
MinimumSupply loops over a list in the tab minimum voorraad, for each item in the list it loops over the list in the tab voorraadbeheer and calculates the total supplies present for the item in minimum voorraad.
The calculation is done by subtracting the used items from the total items. Next the item is compared to a list of minimum item quantities, if the item is lower than the minimum than the te bestellen page and the besteld page are checked to see if the item is present in the first 30 lines. If the item is present then update the current supplies, if not present search the first empty line and add the item name, the active supply and the current date.
itemPresentInList takes the name of the tab and the name of and item and the column number. It will then check if it can find the name in the specified column.

returns an array: alreadyInList is true when the item is present, false if not. rowWithItem: is 0 when not found, otherwise it will be the row number where the name was found.